==> notebooks/1_SetUp.ipynb <==
# Project Recommender System
Our project for Group 6 is regarding Video Games rating. The objective would be to predict the rating based on the different qualifications from the game.
This notebook is the starting point of this project and is the data cleaning and exploration.

### Outline of the Notebook

- Step 1: Setting Up the Environment
- Step 2: Data Cleaning
    - 2.1 Missing Values and Duplicates
    - 2.2 Fix Data Names and Types
- Step 3: Data Description
- Step 4: Save Data

## Step 1: Setting up the Environment

# Libraries for data manipulation
import pandas as pd
import os

# Load and display the dataset

# Path to the dataset, to be changed according to your local setup
PATH = "/Users/agathecauhape/EMLyon 2024-25/Canada/Recommender System/projet/data/"

file_path = os.path.join(PATH, "video_game_reviews.csv")
df = pd.read_csv(file_path)
print(df.head(3))


df.info()

This dataset contains 47774 games rated. It has a total of 18 columns meaning we have many information regarding each games. A lot of them seems to be object so we might have to do a lot of data manipulation and also on column names which are not respecting conventionnal writing. 

Below is a description of each variables:
- Game Title: Title of the game.
- User Rating: Rating of this game.
- Age Group Target: Age Group that is supposed to be the most attracted to the game.
- Price: Price of the game.
- Platform: Platform on what the game can be played.
- Requires Special Device: If the game requires a special installation.
- Developer: Name of the developer.
- Publisher: Name of the publisher.
- Release Year: Release year of the game.
- Genre: Genre of the game.
- Multiplayer: If the game is multiplayer or not.
- Game Length: How long on average it takes to finish the game. 
- Graphics Quality: Quality of graphics.
- Soundtrack Quality: Quality of soundtracks.
- Store Quality: Quality of the store.
- User Review Text: Text review of the user, will be used in either sentiment analysis or RAG analysis. 
- Game Mode: If the game is played online with other players or in standalone.
- Min Number of Players: Minimum players required to play.


# Save df into df_copy for further processing and not to modify the original df
df_copy = df.copy()

## Step 2: Data Cleaning

### 2.1 Missing Values and Duplicates

We need to check the data for missing values and duplicates. If too much appear it can affect the results of modeling anything. 

df_copy.isna().sum()

# Count duplicates
duplicate_count = df_copy.duplicated().sum()
print(f"Number of duplicate rows: {duplicate_count}")

# Get percentage of duplicates
duplicate_percentage = (duplicate_count / len(df_copy)) * 100
print(f"Percentage of duplicates: {duplicate_percentage:.2f}%")

This dataset has no missing values and neither duplicates so no further cleaning is required regarding this. 

### 2.2 Fix Data Names and Types

The columns names are not respecting the python writing conventions and should be in snakecase for easier manipulation hence we simply change them.

# Change the column names
df_copy.columns = [
    'game_title',
    'user_rating',
    'age_group_targeted',
    'price',
    'platform',
    'requires_special_device',
    'developer',
    'publisher',
    'release_year',
    'genre',
    'multiplayer',
    'game_length_hours',
    'graphics_quality',
    'soundtrack_quality',
    'story_quality',
    'user_review_text',
    'game_mode',
    'min_number_of_players'
]


# Check the new column names
df_copy.info()

Some columns would be better as string variables rather than object and others as category therefore we convert them for later analysis.

# Fix the types of the columns
df_copy['game_title'] = df_copy['game_title'].astype('string')
df_copy['user_rating'] = df_copy['user_rating'].astype('float')
df_copy['age_group_targeted'] = df_copy['age_group_targeted'].astype('category')
df_copy['price'] = df_copy['price'].astype('float')
df_copy['platform'] = df_copy['platform'].astype('category')
df_copy['requires_special_device'] = df_copy['requires_special_device'].astype('string')
df_copy['developer'] = df_copy['developer'].astype('string')
df_copy['publisher'] = df_copy['publisher'].astype('string')
df_copy['release_year'] = df_copy['release_year'].astype('int')
df_copy['genre'] = df_copy['genre'].astype('category')
df_copy['multiplayer'] = df_copy['multiplayer'].astype('string')
df_copy['game_length_hours'] = df_copy['game_length_hours'].astype('float')
df_copy['graphics_quality'] = df_copy['graphics_quality'].astype('category')
df_copy['soundtrack_quality'] = df_copy['soundtrack_quality'].astype('category')
df_copy['story_quality'] = df_copy['story_quality'].astype('category')
df_copy['user_review_text'] = df_copy['user_review_text'].astype('string')
df_copy['game_mode'] = df_copy['game_mode'].astype('string')
df_copy['min_number_of_players'] = df_copy['min_number_of_players'].astype('int')


## Step 3: Data Description

Before going into detailed analysis, we want to get an overview of the data to know where to go from there.

df_copy.describe()

# Display the count of unique values for each colmn
print(" Unique value counts per column:\n")
for col in df_copy.columns:
    print(f"{col}: {df_copy[col].nunique()} unique values")

## Step 4: Save Data
We save the dataset df_copy into a CSV file to be able to reuse it in the future of the project this way. 

# Verify the changes made to the DataFrame before saving
df_copy.head(3)

# Save
df_copy.to_csv(os.path.join(PATH, "video_game_clean.csv"), index=False)
